Sprint 5

+++ src/main/java/com/mitrais/chipper/temankondangan/backendapps/service/impl/EventServiceImpl.java
@@ -368,6 +368,10 @@ public class EventServiceImpl implements EventService {
 			throw new BadRequestException("Error: You cannot apply to your own event!");
 		}
 
+		if(Boolean.TRUE.equals(event.getCancelled())) {
+			throw new BadRequestException("Error: You cannot applied to cancelled event");
+		}
+
 		if (Boolean.TRUE.equals(applicantRepository.existsByApplicantUserAndEvent(user, event))) {
 			throw new BadRequestException("Error: You have applied to this event");
 		}
@@ -406,6 +410,10 @@ public class EventServiceImpl implements EventService {
 			throw new BadRequestException("Error: You are already rejected. You don't need to cancel it anymore.");
 		}
 
+		if(Boolean.TRUE.equals(event.getCancelled())) {
+			throw new BadRequestException("Error: You cannot cancel to cancelled event");
+		}
+
 		if (isCancelationValid(event.getStartDateTime())) {
 			applicantRepository.delete(applicant);
 		} else {
@@ -432,8 +440,12 @@ public class EventServiceImpl implements EventService {
 			throw new BadRequestException("Error: You already have canceled this event");
 		}
 
-		event.setCancelled(true);
-		eventRepository.save(event);
+		if (isCancelationValid(event.getStartDateTime())) {
+			event.setCancelled(true);
+			eventRepository.save(event);
+		} else {
+			throw new BadRequestException("Error: The event will be started in less than 24 hours");
+		}
 	}
 
 	@Override
@@ -590,15 +602,19 @@ public class EventServiceImpl implements EventService {
 			throw new BadRequestException("Error: startDate and finishDate must be all empty or all filled!");
 		}
 
+		LocalDateTime currentTime = LocalDateTime.now();
 		DateTimeFormatter dfDate = DateTimeFormatter.ofPattern("dd-MM-uuuu").withResolverStyle(ResolverStyle.STRICT);
-		LocalDateTime startDateSearch = LocalDateTime.now();
-		LocalDateTime finishDateSearch = LocalDateTime.now().plusDays(90);
+		LocalDateTime startDateSearch = currentTime;
+		LocalDateTime finishDateSearch = LocalDate.now().plusDays(90).atTime(LocalTime.MAX);
 
 		if ((StringUtils.isNotEmpty(startDate))) {
 			startDateSearch = LocalDate.parse(startDate, dfDate).atStartOfDay();
+			if (startDateSearch.toLocalDate().equals(LocalDate.now())) {
+				startDateSearch = currentTime;
+			}
 			finishDateSearch = LocalDate.parse(finishDate, dfDate).atTime(LocalTime.MAX);
 
-			if (startDateSearch.isBefore(LocalDateTime.now())) {
+			if (startDateSearch.isBefore(currentTime)) {
 				throw new BadRequestException("Error: Date inputted have to be today or after!");
 			}
